% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{rmdreport.generate}
\alias{rmdreport.generate}
\title{Generate HTML output from RMarkdown file}
\usage{
rmdreport.generate(
  input.filename,
  output.filename,
  stylesheet = getOption("markdown.HTML.stylesheet"),
  ...
)
}
\arguments{
\item{input.filename}{R Markdown input file.}

\item{output.filename}{Markdown or HTML output file.  An HTML file
is specified using the .htm, .html, .HTM or .HTML file extension.
When html is specified, a similarly named markdown file is also
generated.
All output files including cache and figures will appear in the
same folder as \code{output.filename}.}

\item{stylesheet}{Argument passed to \code{markdownToHTML()}
(Default: \code{getOption("markdown.HTML.stylesheet")})}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\value{
Returns \code{output.filename}.
}
\description{
Uses the \code{knitr} R package to generate markdown from input
and then the \code{markdown} R package to generate HTML
}
\examples{
## request a table of contents at the beginning of the output
options(markdown.HTML.options=union('toc', getOption("markdown.HTML.options")))
## use the 'style.css' from this R package to format output HTML
styles <- system.file("style.css", package="rmdreport")
options(markdown.HTML.stylesheet=styles)
## generate markdown and HTML output files in the 'reports/' directory
## from toy example
report <- system.file("example.rmd", package="rmdreport")
rmdreport.generate(report, file.path("reports", "report.html"))

}
